feat: fixes audit M1
When the vault is first deployed and there are a low number of shares
issued it is possible for an adversary to perform a donation attack
wherein they transfer rewards directly to the contract create a skew of
precision between the assets within the contract and the shares issued.

There are a number of protection mechanisms to prevent this and this is
resolved in LRDT by burning a number of shares at the time of deployment
to ensure that the precision between shares and assets cannot differ too
greatly. By default this has been set to `1 ether` and this is required
to be available to the constructor at the time of deployment.