fix: carry over remaining vesting time
The ambition of the public `updateVestingSchedule` was both for this to
be autonomous without admin roles as well as ensuring that a vesting
cycle will execute consistently on the same day and time of the week,
every two weeks.

Therefore if the first vesting period starts at 12pm Monday then we
should expect the next one to finish at this time. The purpose of this
is to allow for cron-like automation where the vesting period can be
updated on a regular schedule and the cyles become predictable and
cosistent for reporting.

Prior to this change the vestingPeriodFinish would drift earlier
depending how long this was called before the finish time.