feat: fixes audit L1
Reimpmlement SafeCast for GLRDT. This was initially removed as each use
would have been protected by the maximum supply, or unfeasible to hit in
realistic scenarios (such as a length of an array exceeding a uint32).

To be on the safe side these have been re-implemented as it will also
help those reading the contract to have confidence that the GLRDT
contract has not been meaningfully changed from OZ's ERC20Votes.sol.