feat: fixes audit L5
One thing lead to another and this commit implements a much more
comprehensive fix for the issue raised.

There were a number of linked issues in the current implementation:

- Due to overriding the ERC20._mint function freeAssets was not yet set
  at the time convertToAssets was called, returning the asset amount
  prior to the freeAssets update.
- Users would need to checkpoint their balance again in order for their
  rewards to be reflected.

That last point was the main motivator for this change. Lets say that
Alice deposits 1 ether at a rate of 1.1, giving her voting power of 1.1
ether. After 100 blocks of rewards the rate then becomes 1.2 and Bob
makes a deposit. At block 100 Bob would have a usable checkpoint with a
voting power of 1.2 ether and Alice's voting power would remain at 1.1
ether, forcing Alice to re-checkpoint in order to capture the assets
earned.

We had attempted to avoid this in the original design but due to a
misunderstanding of how Compound governor read the voting checkpoints
this resulted in incorrect handling.

To fix this we now take a different approach and expose a
convertToAssets function that accepts a blockNumber parameter. This uses
the totalSupplyCheckpoints that are updated on each deposit and burn to
convert the user's share balance within the checkpoint to assets at a
specific block. Note here that `checkpoint.assets` is written for both
totalSupplyCheckpoints and userCheckpoints, but is not used to calculate
voting power on the userCheckpoint.

The key difference instead is that the `checkpoint.shares` is passed to
the new convertToAssets function so that all stakers' voting power is
consistent to the block. Summarised:

1. Upon each mint and burn the total supply of shares and assets is
   recorded.
2. The share balance for each user is checkpointed.
3. When calculating voting power the user's share balance is retrieved
   and conveted to assets using the total supply checkpoint at that
   block.

Therefore Alice would not need to re-checkpoint her rewards earned at
block 100 as this is recalculated due to the totalSupplyCheckpoint
created thanks to Bob's deposit.